Use Windows build number instead of ReleaseId wherever possible (#170)
Motivation: ReleaseId is confusing. In Windows 20H2, ReleaseId is 2009 (so it doesn't match human-visible version). Even worse, ReleaseId is *also* 2009 in Windows 21H2 and we actually need to use DisplayName to distinguish between the two.

Depending on how we proceed with fixing #138 and #166, we might completely drop ReleaseId and only operate on build numbers.

This commit is not expected to change ue4-docker behavior in any way (possibly except for 21H1 that is not supported yet anyway).