Migrate from `setup.py` to `pyproject.toml`
`pyproject.toml` is a mainstream (as of today) way to organize Python packages.

There are multiple reasons for that, I'll just leave some links:

* https://peps.python.org/pep-0517/
* https://peps.python.org/pep-0518/
* https://peps.python.org/pep-0621/
* https://blog.ganssle.io/articles/2021/10/setup-py-deprecated.html

I've also took liberty to add myself to authors list in metadata, hopefully 21% of commits justifies this.

Unfortunately, this stuff doesn't properly work with default pip/setuptools shipped with Python 3.6, so it has to be dropped.